requirements.txt
Flask==0.10.1
Flask-SQLAlchemy
Flask-WTF
Flask-Bootstrap
